refactor: :recycle: move non-critical code to library
